Update structure of the record
