+++ sync/src/main/java/com/spbau/news/on/the/map/sync/entity/ArticleBean.java
@@ -3,26 +3,35 @@ package com.spbau.news.on.the.map.sync.entity;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.util.Date;
+
 public class ArticleBean {
   private final int myId;
   private final int myRawId;
   private final int myGeoId;
   private final String myContent;
+  private final Date myDate;
 
   private final LinksBean myLinks;
-  private final PropertiesBean myProperties;
+
+  private final LocationBean myLocation;
+  private final int myCategory;
 
   @JsonCreator
-  public ArticleBean(@JsonProperty("id") int id, @JsonProperty("rawId") int rawId,
-                     @JsonProperty("geoId") int geoId, @JsonProperty("content") String content,
-                     @JsonProperty("links") LinksBean links, @JsonProperty("properties") PropertiesBean properties) {
+  public ArticleBean(@JsonProperty("id") int id, @JsonProperty("raw_id") int rawId,
+                     @JsonProperty("geo_id") int geoId, @JsonProperty("content") String content,
+                     @JsonProperty("publish_date") Date date, @JsonProperty("links") LinksBean links,
+                     @JsonProperty("location") LocationBean loc, @JsonProperty("category") int category) {
     myId = id;
     myGeoId = geoId;
     myRawId = rawId;
     myContent = content;
 
+    myDate = date;
     myLinks = links;
-    myProperties = properties;
+
+    myLocation = loc;
+    myCategory = category;
   }
 
 
@@ -31,12 +40,12 @@ public class ArticleBean {
     return myId;
   }
 
-  @JsonProperty("rawId")
+  @JsonProperty("raw_id")
   public int getRawId() {
     return myRawId;
   }
 
-  @JsonProperty("geoId")
+  @JsonProperty("geo_id")
   public int getGeoId() {
     return myGeoId;
   }
@@ -51,8 +60,18 @@ public class ArticleBean {
     return myLinks;
   }
 
-  @JsonProperty("properties")
-  public PropertiesBean getProperties() {
-    return myProperties;
+  @JsonProperty("publish_date")
+  public Date getDate() {
+    return myDate;
+  }
+
+  @JsonProperty("location")
+  public LocationBean getLocation() {
+    return myLocation;
+  }
+
+  @JsonProperty("category")
+  public int getCategory() {
+    return myCategory;
   }
 }
