+++ sync/src/test/java/com/spbau/news/on/the/map/sync/entity/BaseBeanTestCase.java
@@ -5,6 +5,5 @@ import java.util.Date;
 abstract class BaseBeanTestCase {
   final LinksBean links = new LinksBean("http://vk.com", "http://vk.com/groupId/wall-123123");
   final LocationBean location = new LocationBean(1.1f, 2.4f);
-  final PropertiesBean properties = new PropertiesBean(location, 10, new Date());
-  protected final ArticleBean article = new ArticleBean(1, 2, 3, "putin", links, properties);
+  protected final ArticleBean article = new ArticleBean(1, 2, 3, "putin",  new Date(), links, location, 23);
 }
